% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{convert_Coo}
\alias{convert_Coo}
\title{Convert to coo a whole Coo Momocs' object}
\usage{
convert_Coo(x)
}
\arguments{
\item{x}{any \code{Coo} object from Momocs}
}
\description{
Convert to coo a whole Coo Momocs' object
}
\seealso{
Other convert: \code{\link{convert_Coo1}},
  \code{\link{convert}}
}
