% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-package.R
\docType{package}
\name{coo}
\alias{coo}
\alias{coo-package}
\title{coo}
\description{
A complete toolkit for morphometrics, from data extraction to multivariate analyses.
Most common 2D morphometrics approaches are included:
outlines, open outlines, configurations of landmarks, traditional morphometrics,
and facilities for data preparation, manipulation and visualization
with a consistent grammar throughout.
Momocs allows reproducible, complex morphometric analyses,
paves the way for a pure open-source workflow in R,
and other morphometrics approaches should be easy to plug in,
or develop from, on top of this canvas.
}
\details{
To cite Momocs in publications: \code{citation("Momocs")}.
}
