% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_Coo}
\alias{export_Coo}
\title{Export to a single or separate coo file a whole Coo Momocs' object}
\usage{
export_Coo(x, file, separate = FALSE, force = FALSE, zip = FALSE)
}
\arguments{
\item{x}{any \code{Coo} object from \code{Momocs}}

\item{file}{\code{character} the name of the file(s) to be created}

\item{separate}{\code{logical} whether to create separate coo files}

\item{force}{\code{logical} whether to replace files}

\item{zip}{\code{logical} whether to zip the created file}
}
\description{
Export to a single or separate coo file a whole Coo Momocs' object
}
\examples{
\dontrun{
library(Momocs)
bot \%>\%
   slice(c(1:5, 21:25)) \%>\% coo_sample(12) \%>\%
   mutate(size=coo_centsize(.)) \%>\%
   export_Coo("bot_lite")
}
}
\seealso{
Other export: \code{\link{export_Coo1}},
  \code{\link{export_shp}}
}
