% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_Coo1}
\alias{export_Coo1}
\title{Export to coo from a single shape from Coo Momocs' objects}
\usage{
export_Coo1(x, id)
}
\arguments{
\item{x}{any \code{Coo} object from Momocs}

\item{id}{\code{numeric}, which shape to export}
}
\description{
Export to coo from a single shape from Coo Momocs' objects
}
\seealso{
Other export: \code{\link{export_Coo}}
}
