% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_shp}
\alias{export_shp}
\title{Export to single coo file from a shape}
\usage{
export_shp(x, file)
}
\arguments{
\item{x}{\code{matrix}}

\item{file}{\code{character} the name of the file to be created}
}
\description{
Nothing else than coordinates is supported
}
\seealso{
Other export: \code{\link{export_Coo1}},
  \code{\link{export_Coo}}
}
