% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_coo}
\alias{import_coo}
\title{Import multiple coo}
\usage{
import_coo(x)
}
\arguments{
\item{x}{coo, as paths or a character vector}
}
\description{
Import multiple coo
}
\seealso{
Other import: \code{\link{import_coo1}}
}
