% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_coo1}
\alias{import_coo1}
\title{Import a single coo}
\usage{
import_coo1(x)
}
\arguments{
\item{x}{coo, as path or character vector}
}
\description{
Import a single coo
}
\seealso{
Other import: \code{\link{import_coo}}
}
