% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_mod}
\alias{import_mod}
\title{Import multiple mod}
\usage{
import_mod(x)
}
\arguments{
\item{x}{mod, as paths or a character vector}
}
\description{
Import multiple mod
}
\seealso{
Other import: \code{\link{import_mod1}}
}
