% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_mod1}
\alias{import_mod1}
\title{Import a single mod}
\usage{
import_mod1(x)
}
\arguments{
\item{x}{mod, as path or `character vector}
}
\description{
Import a single mod
}
\seealso{
Other import: \code{\link{import_mod}}
}
