% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect mod file for possible problems}
\usage{
inspect(x)
}
\arguments{
\item{x}{simple or multiple mod, as path(s) or vector of `character}
}
\description{
Not guaranteed.
}
