% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect1}
\alias{inspect1}
\title{Inspect coo file for possible problems}
\usage{
inspect1(x)
}
\arguments{
\item{x}{simple or multiple coo(s), as path or character vector}
}
\description{
Not guaranteed.
}
