% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-package.R
\docType{package}
\name{Modown}
\alias{Modown}
\alias{Modown-package}
\title{Modown}
\description{
A minimalist syntax for morphometrics data
}
\details{
To cite Modown in publications: \code{citation("Modown")}.
}
