% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mod}
\alias{read_mod}
\title{read multiple mod}
\usage{
read_mod(x)
}
\arguments{
\item{x}{mod, as paths or a character vector}
}
\description{
read multiple mod
}
\seealso{
Other read: \code{\link{read_mod1}}
}
