% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mod1}
\alias{read_mod1}
\title{Read a single mod}
\usage{
read_mod1(x)
}
\arguments{
\item{x}{mod, as path or `character vector}
}
\description{
Read a single mod
}
\seealso{
Other read: \code{\link{read_mod}}
}
