% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate coo to other languages}
\usage{
translate(x)
}
\arguments{
\item{x}{simple or multiple coo(s), as path or character vector}
}
\description{
Plop
}
\seealso{
Other translate: \code{\link{translate_Ldk}},
  \code{\link{translate_Opn}}, \code{\link{translate_Out}}
}
