% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_Ldk}
\alias{translate_Ldk}
\title{Translate mod to Momocs' Ldk}
\usage{
translate_Ldk(x)
}
\arguments{
\item{...}{Arguments passed on to \code{translate_Out}
\describe{
  \item{x}{simple or multiple \code{.mod}, as path(s) or a vector of \code{character}}
}}
}
\description{
Turns into
}
\seealso{
Other translate: \code{\link{translate_Opn}},
  \code{\link{translate_Out}}
}
