% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_Opn}
\alias{translate_Opn}
\title{Translate coo to Momocs' Opn}
\usage{
translate_Opn(x)
}
\arguments{
\item{x}{simple or multiple coo(s), as path or character vector}
}
\description{
Turns into
}
\seealso{
Other translate: \code{\link{translate_Ldk}},
  \code{\link{translate_Out}}
}
