% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_Out}
\alias{translate_Out}
\title{Translate mod to Momocs' Out}
\usage{
translate_Out(x)
}
\arguments{
\item{x}{simple or multiple \code{.mod}, as path(s) or a vector of \code{character}}
}
\description{
Turns into
}
\seealso{
Other translate: \code{\link{translate_Ldk}},
  \code{\link{translate_Opn}}
}
