% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{Write to single coo file from a shape}
\usage{
write(x, file, force = FALSE)
}
\arguments{
\item{x}{\code{matrix}}

\item{file}{\code{character} the name of the file to be created}

\item{force}{\code{logical} whether to replace file}
}
\description{
Nothing else than coordinates is supported
}
\seealso{
Other write: \code{\link{write_Coo1}},
  \code{\link{write_Coo}}
}
