% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_Coo}
\alias{write_Coo}
\title{Write to a single or separate coo file a whole Coo Momocs' object}
\usage{
write_Coo(x, folder = "", file, separate = FALSE, force = FALSE,
  zip = FALSE)
}
\arguments{
\item{x}{any \code{Coo} object from \code{Momocs}}

\item{folder}{\code{character} (optionnal) to write
within a folder, existing or not}

\item{file}{\code{character} the name of the file(s) to be created}

\item{separate}{\code{logical} whether to create separate coo files}

\item{force}{\code{logical} whether to replace files without notice}

\item{zip}{\code{logical} whether to zip the created file}
}
\description{
Write to a single or separate coo file a whole Coo Momocs' object
}
\seealso{
Other write: \code{\link{write_Coo1}}
}
