% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_Coo1}
\alias{write_Coo1}
\title{Write to a single mod file a single shape from a Coo Momocs' object}
\usage{
write_Coo1(x, id, file, force = FALSE)
}
\arguments{
\item{x}{any \code{Coo} object from \code{Momocs}}

\item{id}{\code{numeric}, which shape to write}

\item{file}{\code{character} the name of the file to be created}

\item{force}{\code{logical} whether to replace file}
}
\description{
Write to a single mod file a single shape from a Coo Momocs' object
}
\seealso{
Other write: \code{\link{write_Coo}}, \code{\link{write}}
}
