% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert to mod from a single shape}
\usage{
convert(x)
}
\arguments{
\item{x}{\code{matrix} of coordinates,
or \code{data.frame} of covariates,
or a \code{list} with
\code{$coo} (\code{matrix}) and \code{$cov} (\code{data.frame})}
}
\description{
Returns a mod from coordinates, covariates or both.
}
\seealso{
Other convert: \code{\link{convert_Coo1}},
  \code{\link{convert_Coo}}
}
