% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{convert_Coo1}
\alias{convert_Coo1}
\title{Convert to coo from a single shape from Coo Momocs' objects}
\usage{
convert_Coo1(x, id)
}
\arguments{
\item{x}{any \code{Coo} object from Momocs}

\item{id}{\code{numeric}, which shape to convert}
}
\description{
Convert to coo from a single shape from Coo Momocs' objects
}
\seealso{
Other convert: \code{\link{convert_Coo}},
  \code{\link{convert}}
}
