% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_Coo}
\alias{export_Coo}
\title{Export to coo a whole Coo Momocs' object}
\usage{
export_Coo(x)
}
\arguments{
\item{x}{any \code{Coo} object from Momocs}
}
\description{
Export to coo a whole Coo Momocs' object
}
\seealso{
Other export: \code{\link{export_Coo1}}
}
