% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_Coo1}
\alias{export_Coo1}
\title{Export to a single coo file a single shape from Coo Momocs' objects}
\usage{
export_Coo1(x, id, file, force = FALSE)
}
\arguments{
\item{x}{any \code{Coo} object from \code{Momocs}}

\item{id}{\code{numeric}, which shape to export}

\item{file}{\code{character} the name of the file to be created}

\item{force}{\code{logical} whether to replace file}
}
\description{
Export to a single coo file a single shape from Coo Momocs' objects
}
\seealso{
Other export: \code{\link{export_Coo}},
  \code{\link{export_shp}}
}
