% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_Ldk}
\alias{translate_Ldk}
\title{Translate coo to Momocs' Ldk}
\usage{
translate_Ldk(x)
}
\arguments{
\item{x}{simple or multiple coo(s), as path or character vector}
}
\description{
Turns into
}
\seealso{
Other translate: \code{\link{translate_Opn}},
  \code{\link{translate_Out}}
}
