% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_Out}
\alias{translate_Out}
\title{Translate coo to Momocs' Out}
\usage{
translate_Out(x)
}
\arguments{
\item{x}{simple or multiple coo(s), as path or character vector}
}
\description{
Turns into
}
\seealso{
Other translate: \code{\link{translate_Ldk}},
  \code{\link{translate_Opn}}, \code{\link{translate}}
}
