% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_Coo}
\alias{write_Coo}
\title{Write to a single or separate coo file a whole Coo Momocs' object}
\usage{
write_Coo(x, file, separate = FALSE, force = FALSE, zip = FALSE)
}
\arguments{
\item{x}{any \code{Coo} object from \code{Momocs}}

\item{file}{\code{character} the name of the file(s) to be created}

\item{separate}{\code{logical} whether to create separate coo files}

\item{force}{\code{logical} whether to replace files}

\item{zip}{\code{logical} whether to zip the created file}
}
\description{
Write to a single or separate coo file a whole Coo Momocs' object
}
\examples{
\dontrun{
library(Momocs)
bot \%>\%
   slice(c(1:5, 21:25)) \%>\% coo_sample(12) \%>\%
   mutate(size=coo_centsize(.)) \%>\%
   write_Coo("bot_lite")
}
}
\seealso{
Other write: \code{\link{write_Coo1}}, \code{\link{write}}
}
