% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_Coo1}
\alias{write_Coo1}
\title{Write to a single coo file a single shape from Coo Momocs' objects}
\usage{
write_Coo1(x, id, folder = "", file)
}
\arguments{
\item{x}{any \code{Coo} object from \code{Momocs}}

\item{id}{\code{numeric}, which shape to export}

\item{folder}{\code{character} (optionnal) to write
within a folder, existing or not}

\item{file}{\code{character} the name of the file to be created}
}
\description{
Write to a single coo file a single shape from Coo Momocs' objects
}
\seealso{
Other write: \code{\link{write_Coo}}
}
